% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{assignment}
\alias{assignment}
\title{Assigns a boolean based on a threshold on the 
likelihood or mva score for whether the signature
is identified}
\usage{
assignment(df_in, method = "mva", signame = "Signature_3", data = NULL,
  tumor_type = "breast", do_strict = T)
}
\arguments{
\item{df_in}{input data.frame}

\item{method}{'median_catalog' for likelihood based
selection or 'mva' for multivariate analysis score 
based selection}

\item{signame}{name of the signature that user wants 
to identify, 'Signature_3' or 'Signature_msi'}

\item{data}{'msk', 'seqcap' or 'wgs'}

\item{tumor_type}{tumor type as listed in 
https://github.com/parklab/SigMA/ because the 
thresholds are tumor_type specific}

\item{do_strict}{sets whether a strict 
threshold should be applied or a loose one}
}
\value{
a data.frame with a single column which contains 
the boolean indicating the presence of the signature
}
\description{
Assigns a boolean based on a threshold on the 
likelihood or mva score for whether the signature
is identified
}
