% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setNodeAttr.r
\name{assignPositions}
\alias{assignPositions}
\title{assignPositions}
\usage{
assignPositions(g, positions = layout.random)
}
\arguments{
\item{g}{An igraph object}

\item{positions}{Either a layout function or a layout matrix}
}
\value{
The igraph object with x, y, z coordinates stored in the vertex attributes
}
\description{
Adds position coordinates to an igraph object
}
\details{
Given a graph and either a layout function or a layout matrix, adds those coordiantes to the nodes of the graph.
}
\examples{
data(flightGraph)
assignPositions(flightGraph)
assignPositions(flightGraph, layout_with_fr)
}
\author{
Jared P. Lander
}

