% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkLayout.r
\name{checkLayout}
\alias{checkLayout}
\title{checkLayout}
\usage{
checkLayout(x, layout = layout.random)
}
\arguments{
\item{x}{An igraph object}

\item{Either}{a layout matrix or a function for calculating a graph's layout}
}
\value{
A three-column data.frame with names c('x, 'y', 'z')
}
\description{
Helper function for Sigma to either run a layout function on a graph or accept a prebuilt layout matrix.
}
\details{
If layout is a function it is used on the graph x to calculate a layout matrix.  If layout is a matrix it is used.  Both are converted to a data.frame with x, y, z columns.  This is meant to create a positions object for the igraph.to.rgexf function.
}
\examples{
library(igraph)
data(flightGraph)
data(flightLayout)
checkLayout(flightGraph, layout=layout_with_fr)
checkLayout(flightGraph, layout=flightLayout)
}
\author{
Jared P. Lander
}

