% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cos.R
\name{cosine}
\alias{cosine}
\title{calculates cosine similarity between the spectrum and 
a set of signatures}
\usage{
cosine(x, y)
}
\arguments{
\item{spectrum}{is the mutational spectrum}

\item{signatures}{is the reference signature catalog with
the probability distribution}
}
\description{
calculates cosine similarity between the spectrum and 
a set of signatures
}
