% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\title{Decomposes the mutational spectrum of a genome in terms of
tumor type specific signatures that were calculated through analysis
of public WGS samples from ICGC and TCGA consortia, and contained
as a list in the package. Non-negative-least squares algorithm
is used and the number of signatures to be considered in the 
decomposition is increased gradually, first all pairs from among the
available signatures are considered and minimal error pair is kept. 
Then all 3-signature combinations, 4-signature combinations and so on
are considered. The result is updated if the error is smaller with
larger number of signatures}
\usage{
decompose(spect, signatures, data)
}
\arguments{
\item{spect}{composite spectrum that is being decomposed}

\item{signatures}{a data.frame that contains the signatures in its
columns}

\item{data}{sequencing platform that as in run(), used for setting
the maximum number of signatures that is allowed in the decomposition}
}
\description{
Decomposes the mutational spectrum of a genome in terms of
tumor type specific signatures that were calculated through analysis
of public WGS samples from ICGC and TCGA consortia, and contained
as a list in the package. Non-negative-least squares algorithm
is used and the number of signatures to be considered in the 
decomposition is increased gradually, first all pairs from among the
available signatures are considered and minimal error pair is kept. 
Then all 3-signature combinations, 4-signature combinations and so on
are considered. The result is updated if the error is smaller with
larger number of signatures
}
