% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/igraph.to.json.r
\name{igraphToJson}
\alias{igraphToJson}
\title{igraphToJson}
\usage{
igraphToJson(g)
}
\arguments{
\item{g}{An igraph object}
}
\value{
A string of json text representing the igraph object as an array holding an array each for nodes and edges.
}
\description{
Convert an igraph object to json
}
\details{
Converts an igraph object to json with an array holding an array each for nodes and edges.
}
\examples{
data(flightGraph)
igraph.to.json(flightGraph)
}
\author{
Jared P. Lander
}

