% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lite_df.R
\name{lite_df}
\alias{lite_df}
\title{produces a data.frame with fewer columns for easier use}
\usage{
lite_df(merged_output)
}
\arguments{
\item{merged_output}{is the input data.frame}
}
\description{
produces a data.frame with fewer columns for easier use
}
