% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_matrix.R
\name{make_matrix}
\alias{make_matrix}
\title{Converts an input file list with vcf or maf file paths 
to a matrix, it works for general number of context
and for 1 or 2 strands}
\usage{
make_matrix(directory, file_type = "vcf",
  ref_genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19,
  ncontext = 3, nstrand = 1, chrom_colname = NULL, pos_colname = NULL,
  ref_colname = NULL, alt_colname = NULL)
}
\arguments{
\item{directory}{pointer to the directory where input vcf 
maf files reside}

\item{file_type}{'maf', 'vcf' or 'custom'}

\item{ref_genome}{name of the BSgenome currently set by default to
BSgenome.Hsapiens.UCSC.hg19}

\item{ncontext}{number of bases in the nucleotide sequence which 
makes up the spectrum}

\item{nstrand}{number of strands to be considered, 1 contracts to 
a single strand which for ncontext = 3 gives the commonly used 
96 dimensions}

\item{chrom_colname}{used only for custom files a character 
string defining the colname which holds the chromosome number}

\item{pos_colname}{used only for custom files a character 
string defining the colname which holds the position information}

\item{ref_colname}{used only for custom files a character 
string defining the colname which holds the ref allele}

\item{alt_colname}{used only for custom files a character 
string defining the colname which holds the alt allele}
}
\description{
Converts an input file list with vcf or maf file paths 
to a matrix, it works for general number of context
and for 1 or 2 strands
}
\examples{
by default runs on vcf input and produces 96 dimensional spectra 
make_matrix(directory = 'input')
make_matrix(directory = 'input', 
            file_type = 'vcf',
            ref_genome = BSgenome.Hsapiens.UCSC.hg19,
            ncontext = 5,
            nstrand = 2)
}
