% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_decomposition.R
\name{MC_decomposition}
\alias{MC_decomposition}
\title{Associates each SNV to a specific signature based on 
probabilities. The signatures and their exposures must
be discovered by another algorithm and used as input for
this function}
\usage{
MC_decomposition(genomes, signatures, exposures)
}
\arguments{
\item{genomes}{SNV spectra of the genome.}

\item{signatures}{signatures found in the genome}

\item{exposures}{exposures of the signatures in the .}
}
\value{
A matrix of size ntype x ngenome x nsig
where ntype: SNV spectra dimensions
ngenome: number of genomes
nsig: number of signatures
}
\description{
Associates each SNV to a specific signature based on 
probabilities. The signatures and their exposures must
be discovered by another algorithm and used as input for
this function
}
\examples{
run_decomposition('input_snvs_genome.csv', 
                  'signatures.csv',
                  'exposures.csv')
MC_decomposition(genomes, signatures, exposures)
}
