% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_G_SIGMA.R
\name{plot_all_g_sigmas}
\alias{plot_all_g_sigmas}
\title{Plot all g-sigmas}
\usage{
plot_all_g_sigmas(measure_obj)
}
\arguments{
\item{measure_obj}{the measure-object as the output of sigma_funct}
}
\value{
ggplot plot
}
\description{
plot of all g-sigmas obtained per cluster
}
\examples{
data("out")
plot_all_g_sigmas(out)

}
