% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_sigmas.R
\name{plot_all_sigmas}
\alias{plot_all_sigmas}
\title{Plot all sigmas}
\usage{
plot_all_sigmas(measure_obj)
}
\arguments{
\item{measure_obj}{the measure-object as the output of sigma_funct}
}
\value{
ggplot plot
}
\description{
plot of all sigmas obtained per cluster
}
\examples{
data("out")
plot_all_sigmas(out)

}
