% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_detailed.R
\name{plot_detailed}
\alias{plot_detailed}
\title{Generates a detailed plot per sample}
\usage{
plot_detailed(file = NULL, sample = NULL)
}
\arguments{
\item{file}{the csv file produced by SigMA}

\item{sample}{name to be plotted}
}
\description{
Generates a detailed plot per sample
}
