% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SIGMA.R
\name{plot_sigma}
\alias{plot_sigma}
\title{Plot sigma}
\usage{
plot_sigma(measure_obj)
}
\arguments{
\item{measure_obj}{the measure-object as the output of sigma_funct}
}
\value{
ggplot plot
}
\description{
plot the value of sigma for every cluster
}
\examples{
data("out")
plot_sigma(out)

}
