% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tribase_dist.R
\name{plot_tribase_dist}
\alias{plot_tribase_dist}
\title{plots the 96 dimensional mutational spectrum}
\usage{
plot_tribase_dist(df_snvs, file_name = "test.png", labely = "N SNVs",
  legend = T, text_size = 7)
}
\description{
plots the 96 dimensional mutational spectrum
}
