% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tribase_multipanel.R
\name{plot_tribase_multipanel}
\alias{plot_tribase_multipanel}
\title{This function makes a plot of 96 dimensional context
using a data.frame which has on each of it's columns}
\usage{
plot_tribase_multipanel(matrix_snvs, varnames, signame = "Unknown",
  error.up = rep(0, 96), error.down = rep(0, 96))
}
\description{
This function makes a plot of 96 dimensional context
using a data.frame which has on each of it's columns
}
