% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mva.R
\name{predict_mva}
\alias{predict_mva}
\title{This function uses the trained MVA, in particular gradient 
boosting  models, inside the package to assign a probability
for the existence of the signature of interest.}
\usage{
predict_mva(input, signame, data, tumor_type = "breast")
}
\arguments{
\item{input}{is a data frame that has likelihood cosine 
similarity and total snv values in it's columns}

\item{signame}{name of the signature which is being identified}

\item{data}{determines the sequencing platform see run()}

\item{tumor_type}{tumor type tag see ?run}
}
\value{
a data.frame with a single column with the score
of MVA
}
\description{
This function uses the trained MVA, in particular gradient 
boosting  models, inside the package to assign a probability
for the existence of the signature of interest.
}
