% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sigma.R
\name{renderSigma}
\alias{renderSigma}
\title{A Shiny render function for the sigmajs graph visualization library.}
\source{
sigmajs was developed by ALEXIS JACOMY with the help of  GUILLAUME PLIQUE. See \url{http://sigmajs.org/}
}
\usage{
renderSigma(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates an HTML widget.}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.}
}
\description{
A Shiny render function for the sigmajs graph visualization library.
}

