% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Loads signatures and genome snv matrices and runs the matching}
\usage{
run(genome_file, output_file = "output.csv", sig_catalog = "default",
  custom_sig_file = NULL, rm_sigs = NULL)
}
\arguments{
\item{genome_file}{a file with snv spectra info can be created
from vcf file using @make_genome_matrix() function:
see ?make_genome_matrix}

\item{output_file}{the output file name}

\item{sig_catalog}{an array of 'cosmic' 'pcawg' 'default
'custom', 'default option uses the intersection of cosmic
and pcawg catalogs adding the subsignatures from pcawg of 
existing cosmic signatures, e.g. Signature 7a, 7b, etc, but
the new signatures introduced by PCAWG catalog are not used}

\item{custom_sig_file}{a file with signature distributions}

\item{rm_sigs}{an array with signature names that the user
would like to remove in the decomposition}
}
\description{
Loads signatures and genome snv matrices and runs the matching
}
\examples{
run(genome_file = 'input_genomes.csv', 
    sig_catalog = c('cosmic', 'pcawg'))
run(genome_file = 'input_genomes.csv', 
    sig_catalog = 'custom', 
    custom_sig_file = 'signatures_me.csv')
run(genome_file = 'input_genomes.csv',
    sig_catalog = c('pcawg')
    rm_sigs = 'Signature_40')
}
