% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setNodeAttr.r
\name{setNodeColor}
\alias{setNodeColor}
\title{setNodeColor}
\usage{
setNodeColor(x, node.color = "blue")
}
\arguments{
\item{x}{An igraph object}

\item{node.size}{A character specifying the color}
}
\value{
An igraph object with size set
}
\description{
Set node color if not already set
}
\details{
Checks if node size is set, if not set it
}
\examples{
library(igraph)
data(flightGraph)
flightColor <- setNodeColor(flightGraph)
V(flightColor)$color
flightColor <- setNodeColor(flightColor, 'red')
V(flightColor)$color
}
\author{
Jared P. Lander
}

