% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setNodeAttr.r
\name{setNodeSize}
\alias{setNodeSize}
\title{setNodeSize}
\usage{
setNodeSize(x, node.size = 75)
}
\arguments{
\item{x}{An igraph object}

\item{node.size}{A numeric or integer specifying the size of the nodes}
}
\value{
An igraph object with size set
}
\description{
Set node size if not already set
}
\details{
Checks if node size is set, if not set it
}
\examples{
library(igraph)
data(flightGraph)
flightSize <- setNodeSize(flightGraph)
V(flightSize)$size
flightSize <- setNodeSize(flightSize, 16)
V(flightSize)$size
}
\author{
Jared P. Lander
}

