% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigma}
\alias{sigma}
\alias{sigma.gexf}
\alias{sigma.igraph}
\title{sigma}
\usage{
sigma(x, ...)

\method{sigma}{igraph}(x, layout = layout.random, node.size = 75,
  node.color = "blue", ...)

\method{sigma}{gexf}(x, drawEdges = TRUE, drawNodes = TRUE, width = NULL,
  height = NULL)
}
\arguments{
\item{x}{An igraph object}

\item{layout}{Either an igraph layout function or a two column matrix specifying positions}

\item{node.size}{The default node.size for the nodes}

\item{node.color}{Default color for nodes}

\item{\dots}{Further arguments}
}
\value{
A sigma.js plot
}
\description{
Generates a sigma.js plot of a graph
}
\details{
Generates a sigma.js plot of a graph.  Currently built to handle igraph graphs and gexf files.
}
\author{
Jared P. Lander

Jared P. Lander
}

