% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sigma_json.R
\name{sigma_json_generator}
\alias{sigma_json_generator}
\title{Make JSON-formatted sigma network data}
\usage{
sigma_json_generator(nodes, edges, colorPalette = list(Microbe = "#3288bd",
  Compound = "#ff9d00", Host = "#e31a1c"))
}
\arguments{
\item{nodes}{we need to document the required columns}

\item{edges}{same thing here.}

\item{colorPalette}{Color Palette.}
}
\description{
Just need the node \code{\link{data.frame}}
and the edge \code{\link{data.frame}}.
}

