% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigmaOutput}
\alias{sigmaOutput}
\title{The Shiny widget output thingy}
\usage{
sigmaOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{character string}

\item{width}{duh}

\item{height}{duh}
}
\description{
The Shiny widget output thingy
}

