% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{assignment}
\alias{assignment}
\title{This code assigns a boolean for weather the signature
of interest is identified by applying a threshold on 
the multivariate analysis score or likelihood}
\usage{
assignment(df_in, method = "mva", signame = "Signature_3", data = NULL,
  tumor_type = "breast", do_strict = T)
}
\arguments{
\item{df_in}{input data.frame}

\item{method}{'median_catalog' for likelihood based
selection or 'mva' for multivariate analysis score 
based selection}

\item{signame}{name of the signature that user wants 
to identify, 'Signature_3' or 'Signature_msi'}

\item{tumor_type}{thresholds are specific to
tumor types}

\item{do_strict}{boolean to set whether a strict 
threshold should be applied}
}
\value{
a data.frame with a single column which contains 
the boolean indicating the presence of the signature
}
\description{
This code assigns a boolean for weather the signature
of interest is identified by applying a threshold on 
the multivariate analysis score or likelihood
}
