% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigma}
\alias{sigma}
\title{An htmlwidget wrapper for the sigmajs graph visualization library.}
\source{
sigmajs was developed by  ALEXIS JACOMY with the help of  GUILLAUME PLIQUE. See \url{http://sigmajs.org/}
}
\usage{
sigma(gexf, drawEdges = TRUE, drawNodes = TRUE, width = NULL,
  height = NULL)
}
\arguments{
\item{gexf}{GEXF (Graph Exchange XML Format) data file.}

\item{drawEdges}{logical, whether or not to draw the edges}

\item{drawNodes}{logical, whether or not to draw the nodes}

\item{width}{the width of the graph in pixels}

\item{height}{the height of the graph in pixels}
}
\description{
An htmlwidget wrapper for the sigmajs graph visualization library.
}
\examples{
data <- system.file("examples/ediaspora.gexf.xml", package = "sigma")
sigma(data)
}

