% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigma}
\alias{sigma}
\title{Create the sigma.js object}
\usage{
sigma(jsonData, drawEdges = TRUE, drawNodes = TRUE, width = NULL,
  height = NULL)
}
\arguments{
\item{jsonData}{data}

\item{drawEdges}{logical}

\item{drawNodes}{logical}

\item{width}{integer?}

\item{height}{integer?}
}
\value{
something awesome
}
\description{
Create the sigma.js object
}

