% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sigma_json.R
\name{sigma_json_generator}
\alias{sigma_json_generator}
\title{Make JSON-formatted sigma network data}
\usage{
sigma_json_generator(nodeDataframe, edgeDataframe, palette = list(microbe =
  "#3288bd", compound = "#ff9d00", host = "#e31a1c"))
}
\arguments{
\item{nodeDataframe}{we need to document the required columns}

\item{edgeDataframe}{same thing here.}
}
\description{
Just need the node \code{\link{data.frame}}
and the edge \code{\link{data.frame}}.
}

