% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigmaOutput}
\alias{sigmaOutput}
\title{A Shiny output function for the sigmajs graph visualization library.}
\source{
sigmajs was developed by ALEXIS JACOMY with the help of  GUILLAUME PLIQUE. See \url{http://sigmajs.org/}
}
\usage{
sigmaOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{the output variable to read from.}

\item{width}{the width of the graph in percentage.}

\item{height}{the height of the graph in pixels.}
}
\description{
A Shiny output function for the sigmajs graph visualization library.
}

